fix: sum값 int -> long 변경

+++ src/main/java/calculator/Calcurator.java
@@ -17,7 +17,7 @@ public class Calcurator {
     public void run() {
             String input = messageUtil.requestInput();
             List<Integer> nums = inputParser.parse(input);
-            int sum = computerUtil.getInputInt(nums);
+            long sum = computerUtil.getInputInt(nums);
             messageUtil.printResult(sum);
     }
 }
