+++ src/main/java/calculator/util/ComputerUtil.java
@@ -2,8 +2,8 @@ package calculator.util;
 import java.util.List;
 
 public class ComputerUtil {
-    public static int getInputInt(List<Integer> nums) {
-        int sum = 0;
+    public static long getInputInt(List<Integer> nums) {
+        long sum = 0;
         for (int n : nums) {
             sum += n;
         }
